fix job control ctrl-z suspend program to background and returning to foreground
The curses implementation of signal handlers for signals SIGTSTP and SIGCONT, signals for job control and moving between background and foreground didn't work well because of it refreshing with doupdate, etc., resulting in a glitchy screen, so an own implementation of these signal handlers have been made. These new handlers are quite similar to the original ones.